removing fetchResponse and FetchResponseHeaders as a type
