+++ extensions/amp-a4a/0.1/amp-a4a.js
@@ -1179,7 +1179,7 @@ export class AmpA4A extends AMP.BaseElement {
    * headers. Must be less than or equal to the original size of the ad slot
    * along each dimension. May be overridden by network.
    *
-   * @param {!../../../src/service/xhr-impl.FetchResponseHeaders} responseHeaders
+   * @param {!Headers} responseHeaders
    * @return {?SizeInfoDef}
    */
   extractSize(responseHeaders) {
@@ -1242,7 +1242,7 @@ export class AmpA4A extends AMP.BaseElement {
   /**
    * Send ad request, extract the creative and signature from the response.
    * @param {string} adUrl Request URL to send XHR to.
-   * @return {!Promise<?../../../src/service/xhr-impl.FetchResponse>}
+   * @return {!Promise<?Response>}
    * @protected
    */
   sendXhrRequest(adUrl) {
