+++ extensions/amp-a4a/0.1/signature-verifier.js
@@ -315,7 +315,8 @@ export class SignatureVerifier {
                   'Fast Fetch keyset spec requires Content-Type: ' +
                       'application/jwk-set+json');
               return response.json().then(
-                  jwkSet => {
+                  json => {
+                    const jwkSet = /** @type {!JsonObject} */ (json);
                     // This is supposed to be a JSON Web Key Set, as defined in
                     // Section 5 of RFC 7517. However, the signing service could
                     // misbehave and send an arbitrary JSON value, so we have to
