+++ extensions/amp-ad-network-doubleclick-impl/0.1/amp-ad-network-doubleclick-impl.js
@@ -183,7 +183,7 @@ export class AmpAdNetworkDoubleclickImpl extends AmpA4A {
     /** @protected {boolean} */
     this.useSra = false;
 
-    /** @protected {!Deferred<?../../../src/service/xhr-impl.FetchResponse>} */
+    /** @protected {!Deferred<?Response>} */
     this.sraDeferred = new Deferred();
 
     /** @private {?RefreshManager} */
@@ -683,9 +683,15 @@ export class AmpAdNetworkDoubleclickImpl extends AmpA4A {
     this.ampAnalyticsConfig_ = extractAmpAnalyticsConfig(this, responseHeaders);
     this.qqid_ = responseHeaders.get(QQID_HEADER);
     this.troubleshootData_.creativeId =
-        responseHeaders.get('google-creative-id');
+      /** @type {string} */ (dev().assert(
+          responseHeaders.get('google-creative-id'),
+          'Header google-creative-id is null'
+      ));
     this.troubleshootData_.lineItemId =
-        responseHeaders.get('google-lineitem-id');
+      /** @type {string} */ (dev().assert(
+          responseHeaders.get('google-lineitem-id'),
+          'Header google-lineitem-id is null'
+      ));
     if (this.ampAnalyticsConfig_) {
       // Load amp-analytics extensions
       this.extensions_./*OK*/installExtensionForDoc(
