+++ extensions/amp-form/0.1/form-verifiers.js
@@ -34,7 +34,7 @@ let VerificationErrorDef;
  * Construct the correct form verifier based on whether
  * a config block is present.
  * @param {!HTMLFormElement} form
- * @param {function():Promise<!../../../src/service/xhr-impl.FetchResponse>} xhr
+ * @param {function():Promise<!Response>} xhr
  */
 export function getFormVerifier(form, xhr) {
   if (form.hasAttribute('verify-xhr')) {
@@ -151,7 +151,7 @@ export class DefaultVerifier extends FormVerifier { }
 export class AsyncVerifier extends FormVerifier {
   /**
    * @param {!HTMLFormElement} form
-   * @param {function():Promise<!../../../src/service/xhr-impl.FetchResponse>} xhr
+   * @param {function():Promise<!Response>} xhr
    */
   constructor(form, xhr) {
     super(form);
